Extract log related logics from states
